requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
pandas>=2.2.0
numpy>=1.26.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
